fix(api): dos fallos mas del mismo tipo, y un barrido que los caza
El 500 de Disciplinas tenia hermanos. Se saco todo el SQL que generan los
repositorios con el pool fingido (380 consultas, 187 distintas) y se paso
por PREPARE contra un Postgres de usar y tirar con las migraciones de
db/migrations aplicadas. Dos fallos reales:

- estudiantes.contarEstudiantes: el recorte dejaba $6 y $7 sin aparecer en
  el texto, y un marcador sin usar no tiene tipo que deducir ("no se pudo
  determinar el tipo del parametro $6"). El grado pasa a numerarse aparte
  en esa consulta.
- evaluaciones.marcarEstadoPendiente: el mismo $2 se usaba contra est_id
  (smallint) y contra un literal entero, "tipos de dato inconsistentes".
  Habria reventado al guardar una evaluacion, en la Fase 12.

sql-parametros.test.ts recorre todos los repositorios —mas alcanceDe,
auditar y las diez definiciones de reportes— y comprueba las dos reglas
que ninguna prueba de servicio ve: tantos valores como marcadores, y sin
huecos. Sustituye a conteos-parametros.test.ts, que solo miraba cinco
modulos. Con PVCAR_VOLCAR_SQL=1 vuelca el SQL para repetir el barrido
completo; esta contado en docs/barrido-sql.md.

Tablas y columnas: las 187 consultas preparan limpias, ninguna nombra algo
que no exista en el esquema.

+++ backend/src/modules/evaluaciones/evaluaciones.repository.ts
@@ -858,9 +858,10 @@ export async function marcarEstadoPendiente(
 ): Promise<void> {
   await client.query(
     `UPDATE public.evaluacion_nino_pendiente
-        SET est_id = $2,
+        SET est_id = $2::smallint,
             usu_id_registrador = $3,
-            evaninopen_fecha_finalizacion = CASE WHEN $2 = ${ESTADO.EVALUADO} THEN now() ELSE NULL END
+            evaninopen_fecha_finalizacion = CASE WHEN $2::smallint = ${ESTADO.EVALUADO}
+                                                 THEN now() ELSE NULL END
       WHERE evaninopen_id = $1`,
     [evaninopenId, evaluado ? ESTADO.EVALUADO : ESTADO.PENDIENTE, evaluado ? registrador : null],
   );
